requirements.txt
Flask==1.0.2
numpy==1.16.4
WTForms==2.2.1
passlib==1.7.1
requests==2.21.0
PyYAML==5.1.1
Flask-Mail==0.9.1
pyjwt==2.0.1
Click==7.0
python-decouple

#asynchronous streaming

websocket-client
Flask-Sockets==0.2.1

tornado==5.1.1

hug 
nexmo
#distributed processing
celery
tinydb 
